Add "locking" lines to rects
Lines will be stuck to rects when you resize/move the rects.

There are no test for that functionality so far, will probably revisit
it soon (need to look into maybe having specific points in items, where
you may lock lines to, that would make it easier to implement locking
for all shapes, not only for rects).