Revisit locking
From now on, connecting is possible only to "outlets" of specific items.
Items can define which "outlets" and which "connectors" they have. Any
item with non-empty "connectors" can be connected to any item with
non-empty "outlets". Outlets and Connectors are just relative points
(in the range P[0 0]..P[1 1]).

For now, only :line and :rect-line can be 'connectors' (by P[0 0] and
P[1 1]) and then can be connected only to :rect (with 16 'outlets')